Move dataclass from models to locus and crossmapper.
